+++ bluetooth-traffic/src/test/java/com/idmsuedtirol/bluetoothtrafficelaboration/ElaborationIT.java
@@ -3,15 +3,10 @@ package com.idmsuedtirol.bluetoothtrafficelaboration;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.net.URL;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Properties;
-
-import io.github.cdimascio.dotenv.Dotenv;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -21,13 +16,13 @@ import com.idmsuedtirol.bluetoothtrafficelaboration.ElaborationsInfo.TaskInfo;
 public class ElaborationIT {
 	private final Properties props = new Properties();
 	private DatabaseHelper databaseHelper;
+
 	@Before
 	public void setup() throws IOException, ClassNotFoundException {
-		Dotenv dotenv = Dotenv.load();
-		String jdbcUrl =  dotenv.get("JDBC_URL");
-		  databaseHelper = new DatabaseHelper(jdbcUrl);
-
+		String jdbcUrl = System.getenv("JDBC_URL");
+		databaseHelper = new DatabaseHelper(jdbcUrl);
 	}
+
 	@Test
 	public void testCountBluetooth() throws ClassNotFoundException, IOException, SQLException {
 		ArrayList<TaskInfo> newSelectTaskInfo = databaseHelper.newSelectTaskInfo();